Upgraded project to Visual Studio 2010.
Added Korea 1 default profile.

Added Guid to PatcherSettings class, and added individual Guids to the default profiles. These Guids will be used to identify profiles if the name has been changed so that settings can still found and updated if need be.

Added infrastructure to download profiles from the web in the future.

Updated the version.