Close meridian.exe if open while patching, handle failed dl.
Patcher will close meridian.exe before patching and if any file
downloads fail, it will attempt to close meridian.exe again and
redownload the file once. If it fails again, the local file hash will be
added to the downloaded patch hashes which is saved after patching
completes.

Added a "patch failed" message at the end which is displayed if any
files were not successfully downloaded.