Save path from old profiles if set to default.
Users may have two profiles for 103 - an old one with null Guid and a
new one with the correct Guid. If they are using the old one as the
updated profile still (i.e. is set to default), this path needs to be
saved in the new profile (and the new one set to default so it
automatically gets used).