Add .NET dependency check for Ogre, code cleanup.
- Added a dependency check for .NET version 4.5.1 or higher when
downloading Ogre client. If the user has a lower version, the download
is stopped and the user is directed to download .NET in the patcher's
browser window.
- Event handlers are now removed from a ClientPatcher object when a new
one is instantiated. Not doing this prevents the old one from being
deleted, and all event handlers will be called if the event occurs.
- Moved more code into regions for ease of editing.
- Renamed button1 to btnQuit.
- Added a GetClientPatcher() function to ClientPatchForm.cs which uses
the ClientType of the passed PatcherSettings to decide which
ClientPatcher object type to use. Adds/removes event handlers.