Scan full file for MD5 hash to pick up all Ogre changes.
Scanning the first 1024 bytes was causing some Ogre file changes not to
be detected (e.g. alterations in a large text file that don't alter
length).

The whole file is now hashed, and the filename salting has been removed.
This is the same method used in the new Ogre updater, and is as fast or
faster than the existing method of scanning (code used is directly from
the Ogre updater).